% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/privreg_local.R
\name{privreg_local}
\alias{privreg_local}
\title{Local vertically partitioned data regression}
\usage{
privreg_local(y, Xa, Xb, family = gaussian(), se = TRUE, tol = 1e-12,
  maxit = 10000, debug = TRUE)
}
\arguments{
\item{y}{outcome variable}

\item{Xa}{alice model matrix}

\item{Xb}{bob model matrix}

\item{family}{response family (use family object!)}

\item{se}{whether to compute the standard error}

\item{tol}{tolerance}

\item{maxit}{maximum iterations}

\item{debug}{print debug information}
}
\description{
Perform privreg locally with two vertically partitioned datasets
}
