% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/privreg.R
\docType{data}
\name{PrivReg}
\alias{PrivReg}
\title{Private regression with vertically partitioned data}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
PrivReg
}
\description{
This class sets up the connection and performs estimation for private regression.
}
\keyword{datasets}
